lxd/device: Extract shared CDI helpers into gpu_cdi.go
Move the CDI functions out of `gpuPhysical` struct methods and make them
free functions, so they can be shared with other GPU device types
that support CDI passthrough. A follow-up commit migrates the "mig"
gputype to use these helpers.

Changes:

 - The previous `gpuPhysical` methods become free functions taking a
   `*deviceCommon` argument so any GPU device type can call them without
   duplicating the implementation.

 - Add `applyCDIDeviceToContainer()` that wraps the full per-device setup:
   generate the CDI spec, start the unix-char and bind-mount devices,
   write the hooks file, and register the post-start hook that
   applies the CDI hooks to the container.

 - `postStopCDIDevice()` gains `allowMissingFiles` parameter so callers
   that may not have CDI metadata on disk (e.g. instances started
   before a CDI migration) can request lenient cleanup.
   `gpuPhysical` still requests strict cleanup since CDI is
   the only way it ever writes those files.
