+++ lxd/device/gpu_physical.go
@@ -1,10 +1,8 @@
 package device
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io/fs"
 	"net/http"
 	"os"
 	"path/filepath"
@@ -16,11 +14,9 @@ import (
 	"github.com/canonical/lxd/lxd/device/cdi"
 	deviceConfig "github.com/canonical/lxd/lxd/device/config"
 	pcidev "github.com/canonical/lxd/lxd/device/pci"
-	"github.com/canonical/lxd/lxd/idmap"
 	"github.com/canonical/lxd/lxd/instance"
 	"github.com/canonical/lxd/lxd/instance/instancetype"
 	"github.com/canonical/lxd/lxd/resources"
-	"github.com/canonical/lxd/lxd/storage/filesystem"
 	"github.com/canonical/lxd/lxd/util"
 	"github.com/canonical/lxd/shared"
 	"github.com/canonical/lxd/shared/api"
@@ -130,229 +126,6 @@ func (d *gpuPhysical) Start() (*deviceConfig.RunConfig, error) {
 	return d.startContainer()
 }
 
-// startCDIDevices starts all the devices given in a CDI specification:
-// * `unix-char` (representing the card and non-card devices)
-// * `disk` (representing the mounts)).
-func (d *gpuPhysical) startCDIDevices(configDevices cdi.ConfigDevices, runConf *deviceConfig.RunConfig) error {
-	srcFDHandlers := make([]*os.File, 0)
-	defer func() {
-		for _, f := range srcFDHandlers {
-			_ = f.Close()
-		}
-	}()
-
-	hooksFilePath := d.generateCDIHooksFilePath()
-	deviceConfigFilePath := d.generateCDIConfigDevicesFilePath()
-	devicesPath := d.inst.DevicesPath()
-
-	// Check if there are any remaining CDI devices in the instance devices directory.
-	// If there are, we need to remove them. These can be present in the case where the device stop hook was not called
-	// (e.g. due to an abrupt host shutdown).
-	err := filepath.WalkDir(devicesPath, func(path string, e fs.DirEntry, err error) error {
-		if err != nil {
-			return err
-		}
-
-		if e.IsDir() {
-			return nil
-		}
-
-		// Remove the CDI device files (both unix-char and disk devices as long as JSON CDI metadata files).
-		if strings.HasPrefix(e.Name(), filesystem.PathNameEncode(cdi.CDIUnixPrefix+"."+d.name+".")) ||
-			strings.HasPrefix(e.Name(), filesystem.PathNameEncode(cdi.CDIDiskPrefix+"."+d.name+".")) ||
-			path == hooksFilePath || path == deviceConfigFilePath {
-			err = os.Remove(path)
-			if err != nil {
-				return err
-			}
-		}
-
-		return nil
-	})
-	if err != nil {
-		return err
-	}
-
-	for _, conf := range configDevices.UnixCharDevs {
-		if conf["source"] == "" {
-			return fmt.Errorf("The source of the unix-char device %v used for CDI is empty", conf)
-		}
-
-		if conf["major"] == "" || conf["minor"] == "" {
-			return fmt.Errorf("The major or minor of the unix-char device %v used for CDI is empty", conf)
-		}
-
-		major, err := strconv.ParseUint(conf["major"], 10, 32)
-		if err != nil {
-			return fmt.Errorf("Failed parsing major number %q when starting CDI device: %w", conf["major"], err)
-		}
-
-		minor, err := strconv.ParseUint(conf["minor"], 10, 32)
-		if err != nil {
-			return fmt.Errorf("Failed parsing minor number %q when starting CDI device: %w", conf["minor"], err)
-		}
-
-		uid := conf["uid"]
-		if uid != "" {
-			d.config["uid"] = uid
-		}
-
-		gid := conf["gid"]
-		if gid != "" {
-			d.config["gid"] = gid
-		}
-
-		// Here putting a `cdi.CDIUnixPrefix` prefix with 'd.name' as a device name will create an directory entry like:
-		// <lxd_var_path>/devices/<instance_name>/<cdi.CDIUnixPrefix>.<gpu_device_name>.<path_encoded_relative_dest_path>
-		// 'unixDeviceSetupCharNum' is already checking for dupe entries so we have no validation to do here.
-		err = unixDeviceSetupCharNum(d.state, devicesPath, cdi.CDIUnixPrefix, d.name, conf, uint32(major), uint32(minor), conf["path"], false, runConf)
-		if err != nil {
-			return err
-		}
-	}
-
-	// Create the devices directory if missing.
-	err = os.Mkdir(devicesPath, 0711)
-	if err != nil && !errors.Is(err, fs.ErrExist) {
-		return err
-	}
-
-	for _, conf := range configDevices.BindMounts {
-		if conf["source"] == "" {
-			return fmt.Errorf("The source of the disk device %v used for CDI is empty", conf)
-		}
-
-		srcPath := shared.HostPath(conf["source"])
-		destPath := conf["path"]
-		relativeDestPath := strings.TrimPrefix(destPath, "/")
-
-		// Deduplicate CDI bind mounts by target path across devices.
-		// Multiple CDI GPU devices can require the same runtime files (e.g. /run/nvidia-persistenced/socket).
-		// Only create one mount entry per target path to avoid duplicate lxc.mount.entry conflicts.
-		duplicate := false
-		dents, err := os.ReadDir(devicesPath)
-		if err == nil {
-			for _, e := range dents {
-				decoded := filesystem.PathNameDecode(e.Name())
-				if strings.HasPrefix(decoded, cdi.CDIDiskPrefix+".") && strings.HasSuffix(decoded, "."+relativeDestPath) {
-					duplicate = true
-					break
-				}
-			}
-		}
-
-		if duplicate {
-			continue
-		}
-
-		// This time, the created path will be like:
-		// <lxd_var_path>/devices/<instance_name>/<cdi.CDIDiskPrefix>.<gpu_device_name>.<path_encoded_relative_dest_path>
-		deviceName := filesystem.PathNameEncode(deviceJoinPath(cdi.CDIDiskPrefix, d.name, relativeDestPath))
-		devPath := filepath.Join(devicesPath, deviceName)
-
-		ownerShift := deviceConfig.MountOwnerShiftNone
-		if idmap.CanIdmapMount(devPath, "") {
-			ownerShift = deviceConfig.MountOwnerShiftDynamic
-		}
-
-		options := []string{"bind"}
-		mntOptions := shared.SplitNTrimSpace(conf["raw.mount.options"], ",", -1, true)
-		fsName := "none"
-
-		fileInfo, err := os.Stat(srcPath)
-		if err != nil {
-			return fmt.Errorf("Failed accessing source path %q: %w", srcPath, err)
-		}
-
-		fileMode := fileInfo.Mode()
-		isFile := false
-		if !fileMode.IsDir() {
-			isFile = true
-		}
-
-		f, err := os.OpenFile(srcPath, unix.O_PATH|unix.O_CLOEXEC, 0)
-		if err != nil {
-			return fmt.Errorf("Failed opening source path %q: %w", srcPath, err)
-		}
-
-		srcPath = fmt.Sprintf("/proc/self/fd/%d", f.Fd())
-		srcFDHandlers = append(srcFDHandlers, f)
-
-		// Clean any existing entry.
-		err = os.Remove(devPath)
-		if err != nil && !errors.Is(err, fs.ErrNotExist) {
-			return err
-		}
-
-		// Create the mount point.
-		if isFile {
-			f, err := os.Create(devPath)
-			if err != nil {
-				return err
-			}
-
-			srcFDHandlers = append(srcFDHandlers, f)
-		} else {
-			err := os.Mkdir(devPath, 0700)
-			if err != nil {
-				return err
-			}
-		}
-
-		// Mount the fs.
-		err = DiskMount(srcPath, devPath, false, "", mntOptions, fsName)
-		if err != nil {
-			return err
-		}
-
-		if isFile {
-			options = append(options, "create=file")
-		} else {
-			options = append(options, "create=dir")
-		}
-
-		runConf.Mounts = append(runConf.Mounts, deviceConfig.MountEntryItem{
-			DevName:    deviceName,
-			DevSource:  deviceConfig.DevSourcePath{Path: devPath},
-			TargetPath: relativeDestPath,
-			FSType:     "none",
-			Opts:       options,
-			OwnerShift: ownerShift,
-		})
-
-		runConf.PostHooks = append(runConf.PostHooks, func() error {
-			err := unix.Unmount(devPath, unix.MNT_DETACH)
-			if err != nil {
-				return err
-			}
-
-			return nil
-		})
-	}
-
-	// Serialize the config devices inside the devices directory.
-	f, err := os.Create(d.generateCDIConfigDevicesFilePath())
-	if err != nil {
-		return fmt.Errorf("Could not create the CDI config devices file: %w", err)
-	}
-
-	defer f.Close()
-	err = json.NewEncoder(f).Encode(configDevices)
-	if err != nil {
-		return fmt.Errorf("Could not write to the CDI config devices file: %w", err)
-	}
-
-	return nil
-}
-
-func (d *gpuPhysical) generateCDIHooksFilePath() string {
-	return filepath.Join(d.inst.DevicesPath(), d.name+cdi.CDIHooksFileSuffix)
-}
-
-func (d *gpuPhysical) generateCDIConfigDevicesFilePath() string {
-	return filepath.Join(d.inst.DevicesPath(), d.name+cdi.CDIConfigDevicesFileSuffix)
-}
-
 // startContainer detects the requested GPU devices and sets up unix-char devices.
 // Returns RunConfig populated with mount info required to pass the unix-char devices into the container.
 func (d *gpuPhysical) startContainer() (*deviceConfig.RunConfig, error) {
@@ -369,38 +142,10 @@ func (d *gpuPhysical) startContainer() (*deviceConfig.RunConfig, error) {
 				return nil, errors.New(`MIG GPU notation detected for a "physical" gputype device. Choose a "mig" gputype device instead.`)
 			}
 
-			c, ok := d.inst.(instance.Container)
-			if !ok {
-				return nil, fmt.Errorf("Failed casting instance %q to container", d.inst.Name())
-			}
-
-			configDevices, hooks, err := cdi.GenerateFromCDI(d.state.OS.InUbuntuCore(), d.inst, *cdiID)
+			err := applyCDIDeviceToContainer(&d.deviceCommon, *cdiID, &runConf)
 			if err != nil {
 				return nil, err
 			}
-
-			// Start the devices needed by the CDI specification.
-			err = d.startCDIDevices(*configDevices, &runConf)
-			if err != nil {
-				return nil, err
-			}
-
-			// Generate hooks file path for use with ApplyHooksToContainer
-			hooksFile := d.generateCDIHooksFilePath()
-			f, err := os.Create(hooksFile)
-			if err != nil {
-				return nil, fmt.Errorf("Could not create the CDI hooks file: %w", err)
-			}
-
-			defer f.Close()
-			err = json.NewEncoder(f).Encode(hooks)
-			if err != nil {
-				return nil, fmt.Errorf("Could not write to the CDI hooks file: %w", err)
-			}
-
-			runConf.PostHooks = append(runConf.PostHooks, func() error {
-				return cdi.ApplyHooksToContainer(hooksFile, c)
-			})
 
 			return &runConf, nil
 		}
@@ -654,41 +399,6 @@ func (d *gpuPhysical) pciDeviceDriverOverrideIOMMU(pciDev pcidev.Device, driverO
 	return nil
 }
 
-// stopCDIDevices reads the configDevices and remove potential unix device and unmounts disk mounts.
-func (d *gpuPhysical) stopCDIDevices(configDevices cdi.ConfigDevices, runConf *deviceConfig.RunConfig) error {
-	// Remove ALL the underlying unix-char dev entries created when the CDI device started.
-	err := unixDeviceRemove(d.inst.DevicesPath(), cdi.CDIUnixPrefix, d.name, "", runConf)
-	if err != nil {
-		return err
-	}
-
-	for _, conf := range configDevices.BindMounts {
-		relativeDestPath := strings.TrimPrefix(conf["path"], "/")
-		devPath := filepath.Join(d.inst.DevicesPath(), filesystem.PathNameEncode(deviceJoinPath(cdi.CDIDiskPrefix, d.name, relativeDestPath)))
-		runConf.PostHooks = append(runConf.PostHooks, func() error {
-			// Clean any existing device mount entry. Should occur first before custom volume unmounts.
-			err := DiskMountClear(devPath)
-			if err != nil {
-				return err
-			}
-
-			return nil
-		})
-
-		// The disk device doesn't exist do nothing.
-		if !shared.PathExists(devPath) {
-			return nil
-		}
-
-		// Request an unmount of the device inside the instance.
-		runConf.Mounts = append(runConf.Mounts, deviceConfig.MountEntryItem{
-			TargetPath: relativeDestPath,
-		})
-	}
-
-	return nil
-}
-
 // CanHotPlug returns whether the device can be managed whilst the instance is running.
 // Both CDI and classic GPU devices can be hotplugged for containers.
 func (d *gpuPhysical) CanHotPlug() bool {
@@ -709,12 +419,12 @@ func (d *gpuPhysical) Stop() (*deviceConfig.RunConfig, error) {
 	if cdiID != nil {
 		// This is more efficient than GenerateFromCDI as we don't need to re-generate a CDI
 		// specification to parse it again.
-		configDevices, err := cdi.ReloadConfigDevicesFromDisk(d.generateCDIConfigDevicesFilePath())
+		configDevices, err := cdi.ReloadConfigDevicesFromDisk(cdiConfigDevicesFilePath(d.inst.DevicesPath(), d.name))
 		if err != nil {
 			return nil, err
 		}
 
-		err = d.stopCDIDevices(configDevices, &runConf)
+		err = stopCDIDevices(&d.deviceCommon, configDevices, &runConf)
 		if err != nil {
 			return nil, err
 		}
@@ -747,20 +457,9 @@ func (d *gpuPhysical) postStop() error {
 	if d.inst.Type() == instancetype.Container {
 		cdiID, _ := cdi.ToCDI(d.config["id"])
 		if cdiID != nil {
-			err := unixDeviceDeleteFiles(d.state, d.inst.DevicesPath(), cdi.CDIUnixPrefix, d.name, "")
+			err := postStopCDIDevice(&d.deviceCommon, false)
 			if err != nil {
-				return fmt.Errorf("Failed deleting files for CDI device %q: %w", d.name, err)
-			}
-
-			// Also remove the JSON files that were used to store the CDI related information.
-			err = os.Remove(d.generateCDIHooksFilePath())
-			if err != nil {
-				return fmt.Errorf("Failed deleting CDI hooks file for device %q: %w", d.name, err)
-			}
-
-			err = os.Remove(d.generateCDIConfigDevicesFilePath())
-			if err != nil {
-				return fmt.Errorf("Failed deleting CDI paths to conf file for device %q: %w", d.name, err)
+				return err
 			}
 		} else {
 			err := unixDeviceDeleteFiles(d.state, d.inst.DevicesPath(), "unix", d.name, "")
